refactor: 개별, 공용 관리비 폴더 분리
