--- src/main/java/bjs/zangbu/publicdata/service/managecost/ManageCostService.java
@@ -1,9 +1,0 @@
-package bjs.zangbu.publicdata.service.managecost;
-
-import bjs.zangbu.publicdata.dto.managecost.GasCost;
-import bjs.zangbu.publicdata.dto.managecost.HeatCost;
-
-public interface ManageCostService {
-    HeatCost fetchHeatCost(String kaptCode, String searchDate);
-    GasCost fetchGasCost (String kaptCode, String searchDate);
-}
